# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "service for transport renting",
        "title": "Simbir.GO",
        "contact": {},
        "version": "1"
    },
    "host": "localhost:5555",
    "basePath": "/api",
    "paths": {
...
        "/Payment/Hesoyam/{account_id}": {
            "post": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "Enrich balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Enrich balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "account ID",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Account"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/Rent/End/{rent_id}": {
            "post": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "End rent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rent"
                ],
                "summary": "End rent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rent ID",
                        "name": "rent_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Rent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/Rent/MyHistory": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "Get user history",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rent"
                ],
                "summary": "Get user history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Rent"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/Rent/New/{transport_id}": {
            "post": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "New rent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rent"
                ],
                "summary": "New rent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transport ID",
                        "name": "transport_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "rent type",
                        "name": "rentType",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Rent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/Rent/Transport": {
            "get": {
                "description": "Get Transports to rent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rent"
                ],
                "summary": "Get Transports to rent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "latitude",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "longitude",
                        "name": "long",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "radius",
                        "name": "radius",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "transport type",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Transport"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/Rent/TransportHistory/{transport_id}": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "Transport History",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rent"
                ],
                "summary": "Transport History",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trasnport ID",
                        "name": "transport_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Rent"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/Rent/{rent_id}": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "Get rent info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rent"
                ],
                "summary": "Get rent info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rent ID",
                        "name": "rent_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Transport"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Account": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "isAdmin": {
                    "type": "boolean"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.Rent": {
            "type": "object",
            "required": [
                "priceOfUnit",
                "priceType",
                "timeStart",
                "transportId",
                "userId"
            ],
            "properties": {
                "finalPrice": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "priceOfUnit": {
                    "type": "number"
                },
                "priceType": {
                    "type": "string"
                },
                "timeEnd": {
                    "type": "string"
                },
                "timeStart": {
                    "type": "string"
                },
                "transportId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "models.Transport": {
            "type": "object",
            "required": [
                "canBeRented",
                "color",
                "identifier",
                "latitude",
                "longitude",
                "model",
                "ownerId",
                "transportType"
            ],
            "properties": {
                "canBeRented": {
                    "type": "boolean"
                },
                "color": {
                    "type": "string"
                },
                "dayPrice": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "identifier": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "minutePrice": {
                    "type": "number"
                },
                "model": {
                    "type": "string"
                },
                "ownerId": {
                    "type": "string"
                },
                "transportType": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Authorization": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}